refactor(cpu): change memory callback address parameter from uint16_t to uint32_t
Expand address parameter width in memory callback functions to support
24-bit addressing for future 65816 processor implementation while
maintaining backward compatibility.

- Update function type definitions in fam65xx_types.h
- Modify MOS6502 descriptor creation functions
- Update test harnesses to mask addresses for 16-bit compatibility
- Add extended memory support in processor test runner
- Preserve existing 16-bit address space behavior with masking

+++ code/cpp/src/chip/cpu/fam65xx/mos6502.cpp
@@ -150,8 +150,8 @@ static void initialize_mos6502_descriptor() {
 }
 
 fam65xx_chip_descriptor_t* mos6502_create_descriptor(
-    uint8_t (*read_callback)(void* user_data, uint16_t addr, uint8_t bus_state),
-    void (*write_callback)(void* user_data, uint16_t addr, uint8_t data),
+    uint8_t (*read_callback)(void* user_data, uint32_t addr, uint8_t bus_state),
+    void (*write_callback)(void* user_data, uint32_t addr, uint8_t data),
     void* user_data) {
     
     auto* desc = new fam65xx_chip_descriptor_t;
